Move from using processes to threads
This is a part of the interval design.
As virt-who does not need to take advantage of multiple physical cores,
we have decided to move from using processes to threads.
We consider it more maintainable and reliable to get threads to stop
than processes.

Update tests to use threading where necessary

Fix patching to refer to the right Queue class

Update assertions to ensure the proper termination method is called