Implemements a threadsafe datastore
The implemented threadsafe datastore uses maintains a dictionary of keys and
values. Modifying the dictionary is done exclusively post lock aquisition and
prior to lock release. Every item is pickled on the way in and unpickled on
the way out of the underlying dictionary to ensure no two threads share
references to the same objects.

Use pickle for serialization and deserialization

Use cPickle if available, else use pickle
cPickle is upto 1000 times faster, but not always available.