Adds IntervalThread base class and refactors Virt classes
This also adds a new class DestinationThread. The new class is
able to make connections to various destinations and submits
reports gathered from varying sources.

For future reference this is part of the interval design.

Create IntervalThread base class (for virt and destinations) update tests

Add DestinationThread to handle batching and sending of reports to using any Manager

Add tests for DestinationThread