DestinationThreads now send all reports
This step is the last of the mvp of the interval design.
Herein the main thread no longer sends reports to any
destinations, that is now the job of a set of
DestinationThreads. Below are additional changes in this
commit.

Updating tests to account for new design

Separate out oneshot, print mode from regular mode

Removes obsolete tests, adds framework for additional testing

Ensure env is free of all VIRTWHO* variables for subsequent tests

rhsm username and password are optional, add default dest if none found