Adds EnvParser
This adds a new class full of methods and data structures useful
to parse the environment variables to the values that they dictate
internally in virt-who.

This is the first step of decoupling env/cli parsing. It does not
remove the old code from the code base. No paths utilize the new
class as of this commit!