Adds EffectiveConfig, ConfigSection
This adds a new container classes to manage the total configuration of virt-who.
EffectiveConfig is meant to represent the current configuration of virt-who.
It has dictionary-like access to elements, each of which is meant to be a ConfigSection.

A method 'init_config' has been added to the config module. This method is given two
dictionaries: cli_options, env_options. These dictionaries are expected to be the
values read in from the CLI and ENV vars respectively. This method defines how those
options, as well as other sources of virt-who configuration such as the [defaults] section,
are used. This should make it more clear (at least to us) which parameters will be used first.

Each ConfigSection also has dictionary-like access to elements. It tracks it's own
validation state. This is updated as items are modified in the configuration section.
Another key feature of ConfigSections is the validate method. It is meant (in subclasses)
to handle validation of the configuration section. This includes:
- parsing values to those used internally (strings -> booleans etc).
- Ensuring the section has those parameters required for successful use.
- Returning any messages regarding missing values etc for reporting to the user.

Future work will define subclasses that implement what a valid config section looks like
for each virt backend (and destination depending on impl).

This is the first step of decoupling env/cli parsing. It does not
remove the old code from the code base. No paths utilize the new
class as of this commit!

Updates the config module to use a module level logger, updates log init to use config

Updates to validate the global section of the effective config

Updates to fix log tests

Fixes broken tests

Updates test_virt tests to use StubEffectiveConfig

Fixes complex tests and print mode

Refactor to use an EffectiveConfig made of ConfigSections

Rename global configuration option "print_" to "print"

Produce virt backends using EffectiveConfig

This does a number of things. ConfigManager has been renamed to
DestinationToSourceMapper. It is responsible for managing the mapping
of destinations to sources determined based on the EffectiveConfig.
The Executor now uses the sources from the EffectiveConfig to
create the virt threads.

Past this point (largely) what remains is to create the ConfigSections
for each backend, and to remove the Config, GeneralConfig, GlobalConfig
classes.

Update the libvirt backend to use the config as a dict (temporary)

Only use env/cmdline defaults if nothing was specified on the cli / env

Updates some of the config tests to work in the new scheme

Makes EffectiveConfig a singleton, other updates from review

Improve code readability via new function

Implement subclass of ConfigSection for virt backends

* New subclass is named VirtConfigSection
* This class implements most of validation from old
  class `Config` that is common for virtual backends
* Added several unit tests for this new class
  implemented in `test_config_section_virt.py`
* Fixed several unit tests in test_virt.py (filters)

Refactoring of ConfigSection

* Method `validate()` is split to three methods: `_pre_validate()`,
  `_validate()` and `_post_validate()`. 'Post' and 'pre' methods
  are used for steps shared between subclasses. Method `_validate()`
  is intended for own implementation of validation in subclasses.
* Use attribute `invalidate_keys` and removed `to_remove`, because
  the `to_remove` duplicated functionality od `invalidate_keys`.
* Refactored method `_validate_configs()` as generic methods
  `_validate_list()` in `ConfigSection` and I'm reusing it in
  `VirtConfigSection` subclass
* Fixed `update()` method
* Many other small fixes
* Added support for `REQUIRED` options
* Added bunch of unit tests for `ConfigSection`, `GlobalSection`
  and `EffectiveConfig`. Each class has its own file for unit tests.
  (`test_config_section.py`, `test_config_section_global.py and
  `test_effective_config.py`).
* Stop converting list to str in filter_parameters