Adds EffectiveConfig, ConfigSection
This adds a new container classes to manage the total configuration of virt-who.
EffectiveConfig is meant to represent the current configuration of virt-who.
It has dictionary-like access to elements, each of which is meant to be a ConfigSection.

A method 'init_config' has been added to the config module. This method is given two
dictionaries: cli_options, env_options. These dictionaries are expected to be the
values read in from the CLI and ENV vars respectively. This method defines how those
options, as well as other sources of virt-who configuration such as the [defaults] section,
are used. This should make it more clear (at least to us) which parameters will be used first.

Each ConfigSection also has dictionary-like access to elements. It tracks it's own
validation state. This is updated as items are modified in the configuration section.
Another key feature of ConfigSections is the validate method. It is meant (in subclasses)
to handle validation of the configuration section. This includes:
- parsing values to those used internally (strings -> booleans etc).
- Ensuring the section has those parameters required for successful use.
- Returning any messages regarding missing values etc for reporting to the user.

Future work will define subclasses that implement what a valid config section looks like
for each virt backend (and destination depending on impl).

This is the first step of decoupling env/cli parsing. It does not
remove the old code from the code base. No paths utilize the new
class as of this commit!

Updates the config module to use a module level logger, updates log init to use config

Updates to validate the global section of the effective config

Updates to fix log tests

Fixes broken tests

Updates test_virt tests to use StubEffectiveConfig

Fixes complex tests and print mode

Refactor to use an EffectiveConfig made of ConfigSections

Rename global configuration option "print_" to "print"