Produce virt backends using EffectiveConfig
This does a number of things. ConfigManager has been renamed to
DestinationToSourceMapper. It is responsible for managing the mapping
of destinations to sources determined based on the EffectiveConfig.
The Executor now uses the sources from the EffectiveConfig to
create the virt threads.

Past this point (largely) what remains is to create the ConfigSections
for each backend, and to remove the Config, GeneralConfig, GlobalConfig
classes.

Update the libvirt backend to use the config as a dict (temporary)

Only use env/cmdline defaults if nothing was specified on the cli / env

Updates some of the config tests to work in the new scheme

Makes EffectiveConfig a singleton, other updates from review

Improve code readability via new function