Adds FakeVirtConfigSection
This adds a new class that is responsible for validating a 'fake'
backend.

In addition this adds a the "required" keyword argument to the add_key
method of ConfigSection. This argument (if set to True), causes
the ConfigSection to be considered invalid (and hence dropped).

Additional commit notes/messages below (squashed branch):

Invalid configurations are dropped before the default env/cli is added

Removed the requirement for Guest objects to reference Virts

This modifies the Guest __init__ method to expect the value
of CONFIG_TYPE from the Virt class that collected it to be
given instead of passing a reference to the whole Virt class.

This allows many of the methods of the FakeVirt class that are
used to validate the files that it reads to be either static
or class methods. Which, in turn, nicely allows the
FakeVirtConfigSection to validate the "file" parameter by
checking if it the FakeVirt class will be able to read it.

Only override "type" with "virttype" when "type" is not specified
Reset invalid_keys and keys with defaults to default if not specified
Apply destinations specified in add_key
Update esx to use the new "required" kwarg of add_key
Update libvirtd server validation to use qemu:// if "password" is used
Updates tests to account for datastore using deepcopy (instead of pickle)
Only consider "error" messages returned from validation to mean invalid
Removing the -x option from centos7 dockerfile
Only "error" level messages cause keys to be invalid
Also only use the env/cli when there was not an invalid config specified
Only output validation messages that are error level
Execute validation_methods in the order they are added (add_key)
Change order of fake_virt validation methods