Remove the requirement for Guest objects to reference Virts
This modifies the Guest __init__ method to expect the value
of CONFIG_TYPE from the Virt class that collected it to be
given instead of passing a reference to the whole Virt class.

This allows many of the methods of the FakeVirt class that are
used to validate the files that it reads to be either static
or class methods. Which, in turn, nicely allows the
FakeVirtConfigSection to validate the "file" parameter by
checking if it the FakeVirt class will be able to read it.