Further updates to fix breakage from the merge
This updates a number of tests for the subscriptionmanager Manager class
as well as some for DestinationThreads.

The updates for the DestionationThread tests were to use the new mock
config pieces.

The updates for the subscriptionmanager tests (and class) were primarily
changes to the usage of config objects in the class (from the old scheme
to the new).

The GlobalConfigSection was modified to allow the 'print' key to also be
called 'print_'.