handle not running vms
There are several phases that vms can be in. Only in Running phase there is
hostName available. This patch improves how we handle other phases.

There are potential issues with cert verification when using self signed ca.
In this PR we add new param to connect without cert verification and host
validation.

Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>