1981249: Stop using NTLM which requires MD4 (OpenSSL 3.0)
OpenSSL 3.0 dropped support for MD4 in all but it's legacy mode.
This is a minimal change to stop using NTLM and rely only on
Basic auth. Please note that this is a simiar level of security
as the other virtualization providers.

Please see https://bugzilla.redhat.com/show_bug.cgi?id=1981249