spec: drop redundant "release_number" variable
There is already "release" as variable containing the release number, so
use that instead; this makes it possible to hardcode "1" in "Release".

There is a small behaviour change: the version string of virt-who now
will include also the dist-tag, i.e. the trailing part of the release
number (so ".el9", ".fc37", etc).